+++ modules/dev-benchmarks/src/Goals/LegacyWithSpring.lua
@@ -45,11 +45,11 @@ return function()
 	container.Parent = LocalPlayer.PlayerGui
 	local handle = Roact.mount(Roact.createElement(LegacyWithSpring), container)
 
-	local Heartbeat = Otter.__devHeartbeat
+	local AnimationStepSignal = Otter.__devAnimationStepSignal
 
 	measureAndReport(function()
 		for _ = 1, STEP_COUNT do
-			Heartbeat:Fire()
+			AnimationStepSignal:Fire()
 		end
 	end, STEP_COUNT)