+++ modules/otter/src/createSingleMotor.lua
@@ -4,7 +4,7 @@ local Packages = script.Parent.Parent
 local Signal = require(Packages.Signal)
 local createSignal = Signal.createSignal
 
-local Heartbeat = require(script.Parent.Heartbeat)
+local AnimationStepSignal = require(script.Parent.AnimationStepSignal)
 
 local types = require(script.Parent.types)
 type AnimationValue = types.AnimationValue
@@ -65,7 +65,7 @@ function SingleMotor:start()
 		return
 	end
 
-	self.__connection = Heartbeat:Connect(function(dt)
+	self.__connection = AnimationStepSignal:Connect(function(dt)
 		self:step(dt)
 	end)